授权操作

+++ routes/api.js
@@ -11,16 +11,25 @@ var Note = require('../model/note')
 
 // get notes list
 router.get('/notes',(req,res,next)=>{
-    var opts = {raw:true}
+    // var opts = {raw:true}
+    var opts = {where:{id:0}}
+    if(req.session && req.session.user){
+      let userId = req.session.user.id
+      opts = {where:{userId}}
+    }      
     Note.findAll(opts).then(notes=>{
       res.send({status:200,data:notes})
     }).catch(()=>res.send({status:404,errorMsg:'数据库异常'}))
 })
 //修改星星等级
 router.post('/notes/editStar',(req,res,next)=>{
+    if(!req.session || !req.session.user){
+        return res.send({status:400,errorMsg:'请先登录'})
+      }
   var noteId =req.body.id;
+  var userId = req.session.user.id
   var star = req.body.star;
-  Note.update({star},{where:{id:noteId}}).then(lists=>{
+  Note.update({star},{where:{id:noteId,userId}}).then(lists=>{
       res.send({status:200,msg:'更新成功'})
   }).catch(()=>{
       res.send({status:400,errorMsg:'数据库异常或者你没有权限'})
@@ -28,9 +37,13 @@ router.post('/notes/editStar',(req,res,next)=>{
 })
 //修改内容
 router.post('/notes/edit',(req,res,next)=>{
+    if(!req.session || !req.session.user){
+        return res.send({status:400,errorMsg:'请先登录'})
+      }
     var noteId = req.body.id;
     var note = req.body.note;
-    Note.update({text:note},{where:{id:noteId}}).then(lists=>{
+    var userId = req.session.user.id
+    Note.update({text:note},{where:{id:noteId,userId}}).then(lists=>{
        res.send({status:200,msg:'更新成功'})
     }).catch(_=>{
        res.send({status:400,errorMsg:'数据库异常或者你没有权限'})
@@ -38,10 +51,14 @@ router.post('/notes/edit',(req,res,next)=>{
 })
 
 router.post('/notes/finish',(req,res,next)=>{
+    if(!req.session || !req.session.user){
+        return res.send({status:400,errorMsg:'请先登录'})
+      }
     var noteId = req.body.id;
     var finish = req.body.finish;
+    var userId = req.session.user.id
     console.log(noteId,finish,'hello')
-    Note.update({finish},{where:{id:noteId}}).then(lists=>{
+    Note.update({finish},{where:{id:noteId,userId}}).then(lists=>{
        res.send({status:200,msg:'修改成功'})
     }).catch(_=>{
        res.send({status:400,errorMsg:'数据库异常或者你没有权限'})
@@ -49,13 +66,14 @@ router.post('/notes/finish',(req,res,next)=>{
 })
 
 router.post('/notes/add',(req,res,next)=>{
-    if(!req.body.note){
-       return res.send({status:400,errorMsg:'内容不能为空'})
+    if(!req.session || !req.session.user){
+       return res.send({status:400,errorMsg:'请先登录'})
     }
     var note = req.body.note;
+    var userId = req.session.user.id
     var star = req.body.star;
     var finish = req.body.finish;
-    Note.create({text:note,star,finish}).then(()=>{
+    Note.create({text:note,star,finish,userId}).then(()=>{
         res.send({status:200})
     }).catch(()=>{
         res.send({status:400,errorMsg:'数据库异常或者你没有权限'})
@@ -63,8 +81,12 @@ router.post('/notes/add',(req,res,next)=>{
 })
 
 router.post('/notes/delete',(req,res,next)=>{
+    if(!req.session || !req.session.user){
+        return res.send({status:400,errorMsg:'请先登录'})
+      }
     var noteId = req.body.id
-    Note.destroy({where:{id:noteId}}).then(deleteLen=>{
+    var userId = req.session.user.id
+    Note.destroy({where:{id:noteId,userId}}).then(deleteLen=>{
         if(deleteLen === 0){
             return res.send({status:404,errorMsg:'你没有权限'})
         }
